% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{Client}
\alias{Client}
\title{Initializing the SMART Client}
\usage{
Client(url = "https://localhost:5080", id)
}
\arguments{
\item{url}{The URL of FGLab. Defaults to Localhost:5080}

\item{id}{Experiment ID from FGLab. This is provided from the command line}
}
\description{
This function creates an object that contains the URL of FGLab and ID of your experiment
}
\examples{
client = Client(url="https://localhost:5080", id)
}
\keyword{initialization}
