% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{get_argument}
\alias{get_argument}
\title{Get an argument value}
\usage{
get_argument(name)
}
\arguments{
\item{name}{Name of the argument}
}
\description{
Returns the value of an argument for the given run
}
