% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{send_chart}
\alias{send_chart}
\title{Generating time series charts}
\usage{
send_chart(var_names, values)
}
\arguments{
\item{var_names}{?}

\item{values}{?}
}
\description{
Send an arbitrary number of time series graphs to be displayed in FGLab
}
