% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{send_explanation}
\alias{send_explanation}
\title{Sending explanations}
\usage{
send_explanation(explanation, filename = "explanation.png")
}
\arguments{
\item{value}{The string content TODO}
}
\description{
Sends LIME explanation to FGLab
}
