% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{send_metric}
\alias{send_metric}
\title{Sending metrics}
\usage{
send_metric(metric, value, session = client)
}
\arguments{
\item{metric}{The metric name}
}
\description{
Sends an arbitrary metric over PUT request to FGLab.
}
