% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{send_value}
\alias{send_value}
\title{Sending values}
\usage{
send_value(name, value, session = client)
}
\description{
Sends an arbitrary value over PUT request to FGLab. Works similarly to send_metric but the results are displayed differently
}
