% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{vanguard_init}
\alias{vanguard_init}
\title{Connect to FGLab}
\usage{
vanguard_init(url, project_name, experiment_name, parameters, filepath,
  tags = c(), run_locally = FALSE)
}
\arguments{
\item{url}{The URL of FGLab. Default is likely http://localhost:5080.}

\item{project_name}{Name of the project on FGLab.}

\item{experiment_name}{Name of the experiment on FGLab. Must be unique within project.}

\item{parameters}{A named list of parameters with their default values.}

\item{filepath}{Path of the file to be run, relative to working directory.}

\item{tags}{Vector of tags shown on FGLab.}

\item{run_locally}{If TRUE, execute script locally without going through FGLab.}
}
\description{
Initialises a connection to FGLab and creates a new experiment
}
